weekend_rag/requirements.txt
haystack-ai
chroma-haystack
ollama-haystack
elasticsearch-haystack⏎